lee  修改筛选功能

+++ MySuperApp/Class/Other/otherFile/MyButton.h
@@ -28,7 +28,7 @@
 
 
 @property(nonatomic,assign) BOOL isOpenCell;
-@property(nonatomic,assign) int index;
+@property(nonatomic,assign) NSInteger index;
